SvetogorskRP_Bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
flask>=2.0.0